Oops! Indentation!

+++ src/dablooms.h
@@ -36,10 +36,10 @@ typedef struct {
     unsigned int offset;
     unsigned int counts_per_func;
 
-    uint64_t *salts;
+    uint32_t *salts;
     uint32_t *hashes;
 
-	size_t nsalts;
+    size_t nsalts;
     size_t nfuncs;
     size_t size;
     size_t num_bytes;
